refactor: update ground_segmentation namespace and file structure
