+++ perception/ground_segmentation/src/ray_ground_filter/gencolors.hpp
