+++ perception/ground_segmentation/src/scan_ground_filter/node.hpp
