+++ perception/ground_segmentation/test/test_scan_ground_filter.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "ament_index_cpp/get_package_share_directory.hpp"
-#include "ground_segmentation/scan_ground_filter_nodelet.hpp"
+#include "../src/scan_ground_filter/node.hpp"
 #include "tf2_ros/transform_broadcaster.h"
 
 #include <sensor_msgs/point_cloud2_iterator.hpp>
